+++ core/ui/src/main/java/app/kaito_dogi/mybrary/core/ui/navigation/AppNavHost.kt
@@ -17,8 +17,6 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import app.kaito_dogi.mybrary.core.ui.browser.InternalBrowserLauncher
 import app.kaito_dogi.mybrary.core.ui.browser.rememberInternalBrowserLauncher
-import app.kaito_dogi.mybrary.core.ui.navigation.bar.MainNavigationBar
-import app.kaito_dogi.mybrary.core.ui.navigation.bar.NavigationBarDestination
 import app.kaito_dogi.mybrary.core.ui.navigation.route.AppRoute
 import app.kaito_dogi.mybrary.core.ui.navigation.route.MainRoute
